SetElement implementation as its own class
